test: cover cgroup cleanup across flow builds
